Added html files.
The index.html, hg_towers.html, and hg_alliance.html pages were added. The index.html page was populated with the text from the hg_text.txt. The text was put in paragraph tags. The page was split up with header, section, and footer sections. The telephone number and email was set to links. The logo was added to the index.html page. In the hg_towers.html page text was added from hg_text.txt. The text was put into sections to divide the page up. A navigation bar was added at the top of all pages to navigate the site. Reviews were put in block quotes to organize the page. Any other text was put in paragraph tags. An image was added to the top of the page. The hg_alliance.html page is the exact same format as the hg_tower.html. Comments were added across all of the pages.